requirements.txt
#flask
Flask>=3.0.2
gunicorn>=21.2.0
gevent>=24.2.1
grpcio>=1.62.0
flask-restful>=0.3.10
responses>=0.25.0

# test
pytest-mock>=1.10.1
pytest==6.2.4
mock

#fire-store
firebase-admin>=6.4.0


marshmallow>=3.21.0
marshmallow-dataclass[enum, union]
marshmallow-polyfield
dataclasses-jsonschema>=2.16.0



# secrets
google-cloud-secret-manager>=2.18.2
google-cloud-kms>=2.21.2
crcmod>=1.7

# kms related
six==1.16.0

# from gentopia
tenacity
python-dotenv
arxiv
chardet
cchardet
numexpr
googlesearch-python
wolframalpha
numpy
rich
tqdm
prompt-toolkit
regex
attrs
fastapi
geopy
pydantic==1.9
pytest
PyYAML
setuptools
uvicorn
openai==0.28.0
pexpect
gradio_client
pinecone-client
chroma
tiktoken
scholarly
google-cloud-bigquery==3.17.2